implemented single thread model - based on the idea that update and render time fit insdie the timeslot given by the frame rate -> 60fps -> 16ms time to do both update and Renderer.h
removed any docs and methods regarding threading

+++ src/assets/TileMap.cpp
@@ -63,8 +63,8 @@ namespace magique
             LOG_WARNING("Could not load tilemap layer dimensions: %s", asset.path);
             return;
         }
-        width = layerWidth;
-        height = layerHeight;
+        width = static_cast<int16_t>(layerWidth);
+        height = static_cast<int16_t>(layerHeight);
 
         cxstructs::str_skip_char(workPtr, '\n', 2); // Skip to data start
